configure.ac: drop -Wcovered-switch-default flags from llvm-config --cxxflags
Some version of llvm-config include -Wcovered-switch-default in the output
of "llvm-config --cxxflags".  Since this flag is not recognized by gcc,
this prevents the clang headers from being detected when gcc is used
to compile isl.
Arguably, this flag should not be included in the output
of "llvm-config --cxxflags", but it does happen so remove it
from this output.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>