schedule construction: use only positive parameter coefficients
Ever since the original implementation of the isl scheduler in
1192654 (add isl_union_set_compute_schedule, Sat Mar 19 14:47:47 2011 +0100),
the constructed schedules may involve negative coefficients,
both for the variables and the parameters, but only non-negative
constant terms.  It is sufficient to only consider non-negative
constant terms because a schedule involving negative constant terms
can be replaced by an equivalent schedule with only non-negative
constant terms by shifting over the opposite of the most negative
constant term.  The same reasoning also applies to the coefficients
of the parameters, even if the values of the parameters themselves
may be negative, so it is not clear why support for negative
parameter coefficients was introduced in the first place.
Remove it as it can result in counter-intuitive schedules with
negative offsets.  Although removing support for negative parameter
coefficients reduces the total number of variables in schedule (I)LP
problems, it is not clear if there are any scheduling instances where
this results in a significant speed-up.  The effect on the isl test
cases is negligible.

Note that the uniform dependence distance bound should still
be allowed to have negative coefficients in case the parameters
are negative.

Requested-by: Albert Cohen <Albert.Cohen@inria.fr>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>