make schedule_max_coefficient option work as advertised
The schedule_max_coefficient is documented as limiting the values
of the schedule coefficients to the given value.  However, the bound
was not actually being imposed on the schedule coefficients themselves,
but rather on some linear transformation of those coefficients.
The schedule coefficients themselves could therefore in theory end up
exceeding the bound.
Impose the bound on the actual schedule coefficients.
This is probably not all that important in the current state,
but in the future different bounds will be imposed on different coefficients
and in that case it is important that the bounds are imposed on
the correct coefficients.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>